added images and fixed errors

+++ include/gamestate.h
@@ -3,6 +3,6 @@
 
 void resizeBg(char *path, ALLEGRO_BITMAP *background, int width, int height);
 
-bool fadeState(float *visibility, bool closing, int fps);
+bool fadeState(float *visibility, bool closing, float fps);
 
 #endif
